+++ src/ru/khrebtov/lesson7/DepthFirstPath.java
@@ -2,43 +2,18 @@ package ru.khrebtov.lesson7;
 
 import java.util.LinkedList;
 
-public class DepthFirstPath {
-    private boolean[] marked;
-    private int[] edgeTo;
-    private int source;
+public class DepthFirstPath extends BasePath {
 
     public DepthFirstPath(Graph g, int source) {
-        this.source = source;
-        edgeTo = new int[g.getVertexCount()];
-        marked = new boolean[g.getVertexCount()];
-        dfs(g, source);
-    }
-
-    private void dfs(Graph g, int v) {
-        marked[v] = true;
-        for (int w : g.getAdjList(v)) {
-            if (!marked[w]) {
-                edgeTo[w] = v;
-                dfs(g, w);
-            }
-        }
+        super(g, source);
+        super.dfs(g, source);
     }
 
     public boolean hasPathTo(int dist) {
-        return marked[dist];
+        return super.hasPathTo(dist);
     }
 
     public LinkedList<Integer> pathTo(int dist) {
-        if (!hasPathTo(dist)) {
-            return null;
-        }
-        LinkedList<Integer> stack = new LinkedList<>();
-        int vertex = dist;
-        while (vertex != source) {
-            stack.push(vertex);
-            vertex = edgeTo[vertex];
-        }
-        return stack;
+        return super.pathTo(dist);
     }
-
 }
